Revert "AllInoutを追加"
This reverts commit 48922e45a6268b35719ba7046261903cabc0beb1.
