+++ moneybook/templates/_statistics_task_bar.html
@@ -1,6 +1,5 @@
 <div class="task_bar">
     <ul>
         <li><a href="{% url 'moneybook:period_balances' %}">途中残高</a></li>
-        <li><a href="{% url 'moneybook:all_inout' %}">全収支</a></li>
     </ul>
 </div>
