--- moneybook/templates/all_inout.html
@@ -1,92 +1,0 @@
-{% extends "_base.html" %}
-{% load static %}
-{% load humanize %}
-
-{% block title %}{{ app_name }} AllInout{% endblock %}
-
-{% block header %}
-<link rel="stylesheet" type="text/css" href="{% static 'statistics.css' %}">
-<script src="{% static 'statistics_common.js' %}"></script>
-<script src="{% static 'all_inout.js' %}"></script>
-<script src="{% static 'amcharts4/core.js' %}"></script>
-<script src="{% static 'amcharts4/charts.js' %}"></script>
-<script src="{% static 'amcharts4/animated.js' %}"></script>
-<script src="{% static 'amcharts4/kelly.js' %}"></script>
-<script type="text/javascript">
-    all_inout_url = "{% url 'moneybook:all_inout' %}";
-    window.onload = function () {
-        drawGraph();
-    }
-</script>
-{% endblock %}
-
-{% block content %}
-{% include "_statistics_task_bar.html" %}
-<section>
-    <h1>ジャンプ</h1>
-    <table class="tbl-common fcs-blue">
-        <tr>
-            <td rowspan="2">
-                <a href="{% url 'moneybook:all_inout_month' year|add:'-1' %}"><img src="{% static 'images/left.png' %}"
-                        alt="←" width="40px" valign="middle"></a>
-            </td>
-            <td class="jump-date">
-                <input type="text" name="year" id="jump_year" style="width:6ex" class="righter" value="{{ year }}"
-                    onkeypress="keyPressMove(event.keyCode)" onfocus="deleteValue(this)">年
-            </td>
-            <td rowspan="2">
-                <a href="{% url 'moneybook:all_inout_month' year|add:'1' %}"><img src="{% static 'images/right.png' %}"
-                        alt="→" width="40px" valign="middle"></a>
-            </td>
-        </tr>
-        <tr>
-            <td><input type="button" class="btn-apply" value="更新" onclick="moveYear()"></td>
-        </tr>
-    </table>
-    <h1>全収支</h1>
-    <table class="tbl-per-month tbl-common">
-        <tr>
-            <th></th>
-            {% for m in month_all_io_list %}
-            <th>{{ m.label }}</th>
-            {% endfor %}
-        </tr>
-        <tr>
-            <th>収入</th>
-            {% for m in month_all_io_list %}
-            <td>{{ m.income| intcomma }}</td>
-            {% endfor %}
-        </tr>
-        <tr>
-            <th>支出</th>
-            {% for m in month_all_io_list %}
-            <td>{{ m.outgo| intcomma }}</td>
-            {% endfor %}
-        </tr>
-        <tr>
-            <th>収支</th>
-            {% for m in month_all_io_list %}
-            <td>{{ m.balance| intcomma }}</td>
-            {% endfor %}
-        </tr>
-        <tr>
-            <th>途中残高</th>
-            {% for b in period_balances %}
-            <td>{{ b.value |intcomma }}</td>
-            {% endfor %}
-        </tr>
-    </table>
-    <h3>全収入・支出</h3>
-    <div id="barplot_inout_all" class="graph-statistic"></div>
-    <ul id="month_all_io_list" hidden>
-        {% for m in month_all_io_list %}<li>{{ m.label }}月,{{ m.income }},{{ m.outgo }}</li>
-        {% endfor %}
-    </ul>
-    <h3>途中残高</h3>
-    <div id="lineplot_monthly_balance" class="graph-statistic"></div>
-    <ul id="monthly_balance" hidden>
-        {% for b in period_balances %}<li>{{ b.label }}月,{{ b.value }}</li>
-        {% endfor %}
-    </ul>
-</section>
-{% endblock %}
