+++ moneybook/templates/statistics.html
@@ -6,7 +6,6 @@
 
 {% block header %}
 <link rel="stylesheet" type="text/css" href="{% static 'statistics.css' %}">
-<script src="{% static 'statistics_common.js' %}"></script>
 <script src="{% static 'statistics.js' %}"></script>
 <script src="{% static 'amcharts4/core.js' %}"></script>
 <script src="{% static 'amcharts4/charts.js' %}"></script>
